Move window retrieving logic to window manager
