+++ Pinpoint.Plugin.Windower/WindowerPlugin.cs
@@ -1,9 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Linq;
-using System.Runtime.InteropServices;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -19,7 +16,7 @@ public class WindowerPlugin : AbstractPlugin
     public async override IAsyncEnumerable<AbstractQueryResult> ProcessQuery(Query query, CancellationToken ct)
     {
         var app = string.Join(" ", query.Parts[1..]);
-        var windows = GetOpenWindows();
+        var windows = WindowManager.GetWindows();
 
         foreach (var result in windows
             .Where(w => w.QueryMatch(app))
@@ -33,59 +30,4 @@ public class WindowerPlugin : AbstractPlugin
     {
         return Task.FromResult(query.Prefix() == "w" && query.Parts.Length > 1);
     }
-
-    private static List<WindowModel> GetOpenWindows()
-    {
-        var shell = GetShellWindow();
-        var windows = new List<WindowModel>();
-
-        EnumWindows(delegate (IntPtr hWnd, IntPtr lParam)
-        {
-            if (hWnd == shell || !IsWindowVisible(hWnd))
-            {
-                return true;
-            }
-
-            var length = GetWindowTextLength(hWnd);
-            if (length == 0)
-            {
-                return true;
-            }
-
-            var builder = new StringBuilder(length);
-            GetWindowText(hWnd, builder, length + 1);
-
-            GetWindowThreadProcessId(hWnd, out var pid);
-            var process = Process.GetProcessById(pid);
-
-            windows.Add(new WindowModel(builder.ToString(), process.ProcessName, hWnd));
-
-            return true;
-        }, IntPtr.Zero);
-
-        return windows;
-    }
-
-    [DllImport("user32.dll")]
-    private static extern IntPtr GetShellWindow();
-
-    [DllImport("user32.dll")]
-    private static extern bool EnumWindows(EnumWindowsProc enumProc, IntPtr lParam);
-
-    private delegate bool EnumWindowsProc(IntPtr hWnd, IntPtr lParam);
-
-    [DllImport("user32.dll")]
-    private static extern int GetWindowText(IntPtr hWnd, StringBuilder lpString, int nMaxCount);
-
-    [DllImport("user32.dll")]
-    private static extern int GetWindowTextLength(IntPtr hWnd);
-
-    [DllImport("user32.dll")]
-    private static extern bool SetForegroundWindow(IntPtr hWnd);
-
-    [DllImport("user32.dll")]
-    private static extern bool IsWindowVisible(IntPtr hWnd);
-
-    [DllImport("user32.dll")]
-    private static extern uint GetWindowThreadProcessId(IntPtr hWnd, out int lpdwProcessId);
 }
